+++ Sources/GoldSun/Views/EngineHostView.swift
@@ -2,8 +2,16 @@ import SwiftUI
 
 struct EngineHostView: View {
     @ObservedObject var tab: BrowserTabSession
+    @ObservedObject var model: BrowserModel
+    @ObservedObject var downloadStore: DownloadStore
 
     var body: some View {
-        WebKitBrowserView(tab: tab)
+        WebKitBrowserView(
+            tab: tab,
+            downloadStore: downloadStore,
+            openURLInNewTab: { url in
+                model.open(url, inNewTab: true)
+            }
+        )
     }
 }